+++ src/jk_mediawiki/MediaWikiLocalUserServiceMgr.py
@@ -27,6 +27,14 @@ from jk_typing import *
 #
 class MediaWikiLocalUserServiceMgr(object):
 
+	################################################################################################################################
+	## Constants
+	################################################################################################################################
+
+	################################################################################################################################
+	## Constructor
+	################################################################################################################################
+
 	#
 	# Configuration parameters:
 	#
@@ -34,13 +42,16 @@ class MediaWikiLocalUserServiceMgr(object):
 	#										If not specified no shutdown and restart can be performed.
 	# @param	str startPHPFPMScript		The absolute file path of a script that starts an user space PHP process in the background.
 	#										If not specified no shutdown and restart can be performed.
+	# @param	str localEtcDirPath			The path of the local 'etc' directory used by the NGINX and PHP process
 	# @param	str userName				The name of the user account under which NGINX, PHP and the Wiki cron process are executed.
 	#
 	@checkFunctionSignature()
 	def __init__(self,
 		startNGINXScript:str,
 		startPHPFPMScript:str,
+		localEtcDirPath:str,
 		userName:str,
+		bVerbose:bool = False,
 		):
 
 		# store and process the account name the system processes are running under
@@ -60,10 +71,24 @@ class MediaWikiLocalUserServiceMgr(object):
 			assert isinstance(startPHPFPMScript, str)
 			assert os.path.isfile(startPHPFPMScript)
 
+		assert isinstance(localEtcDirPath, str)
+		assert os.path.isdir(localEtcDirPath)
+
 		self.__startNGINXScriptFilePath = startNGINXScript
 		self.__startNGINXScriptDirPath = os.path.dirname(startNGINXScript) if startNGINXScript else None
 		self.__startPHPFPMScriptFilePath = startPHPFPMScript
 		self.__startPHPFPMScriptDirPath = os.path.dirname(startPHPFPMScript) if startPHPFPMScript else None
+		self.__localEtcDirPath = localEtcDirPath
+		self.__bVerbose = bVerbose
+	#
+
+	################################################################################################################################
+	## Public Properties
+	################################################################################################################################
+
+	@property
+	def localEtcDirPath(self) -> str:
+		return self.__localEtcDirPath
 	#
 
 	@property
@@ -86,16 +111,30 @@ class MediaWikiLocalUserServiceMgr(object):
 		return self.__startPHPFPMScriptDirPath
 	#
 
-	def isPHPFPMRunning(self):
-		return self.getPHPFPMMasterProcesses() is not None
+	################################################################################################################################
+	## Helper Methods
+	################################################################################################################################
+
+	################################################################################################################################
+	## Public Methods
+	################################################################################################################################
+
+	def isPHPFPMRunning(self, debugLog:jk_logging.AbstractLogger = None):
+		return self.getPHPFPMMasterProcesses(debugLog) is not None
 	#
 
-	def isNGINXRunning(self):
-		return self.getNGINXMasterProcesses() is not None
+	def isNGINXRunning(self, debugLog:jk_logging.AbstractLogger = None):
+		return self.getNGINXMasterProcesses(debugLog) is not None
 	#
 
-	def stopPHPFPM(self, log = None):
-		processes = self.getPHPFPMMasterProcesses()
+	#
+	# This method stops PHP-FPM processes if they are running.s
+	# On error an exception is raised.
+	#
+	# NOTE: Debug information is written to the log if verbose output is enabled.
+	#
+	def stopPHPFPM(self, log:jk_logging.AbstractLogger):
+		processes = self.getPHPFPMMasterProcesses(log if self.__bVerbose else None)
 		if processes:
 			log.info("Now stopping PHP-FPM processes: " + str([ x["pid"] for x in processes ]))
 			if not jk_utils.processes.killProcesses(processes, log):
@@ -104,8 +143,14 @@ class MediaWikiLocalUserServiceMgr(object):
 			log.notice("No PHP-FPM processes active.")
 	#
 
-	def stopNGINX(self, log = None):
-		processes = self.getNGINXMasterProcesses()
+	#
+	# This method stops NGINX processes if they are running.s
+	# On error an exception is raised.
+	#
+	# NOTE: Debug information is written to the log if verbose output is enabled.
+	#
+	def stopNGINX(self, log:jk_logging.AbstractLogger):
+		processes = self.getNGINXMasterProcesses(log if self.__bVerbose else None)
 		if processes:
 			log.info("Now stopping NGINX processes: " + str([ x["pid"] for x in processes ]))
 			if not jk_utils.processes.killProcesses(processes, log):
@@ -114,51 +159,79 @@ class MediaWikiLocalUserServiceMgr(object):
 			log.notice("No NGINX processes active.")
 	#
 
-	def startPHPFPM(self, log = None):
-		if self.getPHPFPMMasterProcesses() is not None:
+	#
+	# This method starts the PHP-FPM process.
+	# On error an exception is raised.
+	#
+	# NOTE: Debug information is written to the log if verbose output is enabled.
+	#
+	def startPHPFPM(self, log:jk_logging.AbstractLogger):
+		if self.getPHPFPMMasterProcesses(log if self.__bVerbose else None) is not None:
 			raise Exception("PHP-FPM process already running!")
 		if not jk_utils.processes.runProcessAsOtherUser(
 				accountName=self.__userName,
 				filePath=self.__startPHPFPMScriptFilePath,
 				args=None,
-				log=log
+				log=log if self.__bVerbose else None
 			):
 			raise Exception("Starting PHP-FPM process failed!")
+		log.info("PHP-FPM started.")
 	#
 
-	def startNGINX(self, log = None):
-		if self.getNGINXMasterProcesses() is not None:
+	#
+	# This method starts the NGINX process.
+	# On error an exception is raised.
+	#
+	# NOTE: Debug information is written to the log if verbose output is enabled.
+	#
+	def startNGINX(self, log:jk_logging.AbstractLogger):
+		if self.getNGINXMasterProcesses(log if self.__bVerbose else None) is not None:
 			raise Exception("NGINX process already running!")
 		if not jk_utils.processes.runProcessAsOtherUser(
 				accountName=self.__userName,
 				filePath=self.__startNGINXScriptFilePath,
 				args=None,
-				log=log
+				log=log if self.__bVerbose else None
 			):
 			raise Exception("Starting NGINX process failed!")
+		log.info("NGINX started.")
 	#
 
 	#
 	# Returns the master process(es) of "php-fpm". This should be only one process.
 	#
-	def getPHPFPMMasterProcesses(self) -> typing.Union[list, None]:
+	def getPHPFPMMasterProcesses(self, debugLog:jk_logging.AbstractLogger = None) -> typing.Union[list, None]:
 		if self.__startPHPFPMScriptDirPath is None:
 			return None
 
 		processList = jk_sysinfo.get_ps()
 
+		if debugLog and self.__bVerbose:
+			debugLog = debugLog.descend("Scanning for processes ...")
+		else:
+			# NOTE: if no debugging is enabled, no debug logger will be used.
+			debugLog = None
+
 		ret = []
 		for x in processList:
 			if x["user"] != self.__userName:
+				if debugLog:
+					debugLog.debug("Rejecting because not owned by user '{}': {}".format(self.__userName, x))
 				continue
 			if x["cmd"].find("php-fpm") < 0:
-				continue
-			if "cwd" not in x:
-				continue
-			if x["cwd"] != self.__startPHPFPMScriptDirPath:
+				if debugLog:
+					debugLog.debug("Rejecting because command does not contain 'php-fpm': {}".format(x))
 				continue
 			if not x["args"].startswith("master process"):
+				if debugLog:
+					debugLog.debug("Rejecting because 'args' does not start with 'master process': {}".format(self.__startPHPFPMScriptDirPath, x))
 				continue
+			if x["args"].find("(" + self.__localEtcDirPath) < 0:
+				if debugLog:
+					debugLog.debug("Rejecting because does not seem to refer to the local configuration directory: {}".format(x))
+				continue
+			if debugLog:
+				debugLog.debug("Accepting: {}".format(x))
 			ret.append(x)
 
 		return ret if ret else None
@@ -167,28 +240,46 @@ class MediaWikiLocalUserServiceMgr(object):
 	#
 	# Returns the master process(es) of "nginx". This should be only one process.
 	#
-	def getNGINXMasterProcesses(self) -> typing.Union[list, None]:
+	def getNGINXMasterProcesses(self, debugLog:jk_logging.AbstractLogger = None) -> typing.Union[list, None]:
 		if self.__startNGINXScriptDirPath is None:
 			return None
 
 		processList = jk_sysinfo.get_ps()
 
+		if debugLog and self.__bVerbose:
+			debugLog = debugLog.descend("Scanning for processes ...")
+		else:
+			# NOTE: if no debugging is enabled, no debug logger will be used.
+			debugLog = None
+
 		ret = []
 		for x in processList:
 			if x["user"] != self.__userName:
+				if debugLog:
+					debugLog.debug("Rejecting because not owned by user '{}': {}".format(self.__userName, x))
 				continue
 			if not x["cmd"].startswith("nginx"):
-				continue
-			if "cwd" not in x:
-				continue
-			if x["cwd"] != self.__startNGINXScriptDirPath:
+				if debugLog:
+					debugLog.debug("Rejecting because command does not start with 'nginx': {}".format(x))
 				continue
 			if not x["args"].startswith("master process"):
+				if debugLog:
+					debugLog.debug("Rejecting because 'args' does not start with 'master process': {}".format(self.__startPHPFPMScriptDirPath, x))
 				continue
+			if x["args"].find("-c " + self.__localEtcDirPath) < 0:
+				if debugLog:
+					debugLog.debug("Rejecting because does not seem to refer to the local configuration directory: {}".format(x))
+				continue
+			if debugLog:
+				debugLog.debug("Accepting: {}".format(x))
 			ret.append(x)
 
 		return ret if ret else None
 	#
+
+	################################################################################################################################
+	## Public Static Methods
+	################################################################################################################################
 
 #