+++ src/jk_mediawiki/__init__.py
@@ -1,13 +1,14 @@
 ﻿
 
 
-__version__ = "0.2021.1.4"
+__version__ = "0.2021.11.14"
 
 
 
 ################################################################################################################################
 
 from .lang_support_php import PHPTokenizer, PHP
+from .MediaWikiSkinInfo import MediaWikiSkinInfo
 from .MediaWikiLocalSettingsFile import MediaWikiLocalSettingsFile
 from .MediaWikiExtensionInfo import MediaWikiExtensionInfo
 from .MediaWikiDiskUsageInfo import MediaWikiDiskUsageInfo
@@ -17,5 +18,4 @@ from .MediaWikiLocalUserInstallationMgr import MediaWikiLocalUserInstallationMgr
 from .LocalMediaWikisMgr import LocalMediaWikisMgr
 
 ################################################################################################################################
-